Cleanup & Refactor Randomizer code
