+++ Gopkg.toml
@@ -30,6 +30,22 @@
   version = "8.0.6"
 
 [[constraint]]
+  name = "github.com/google/uuid"
+  version = "1.1.1"
+
+[[constraint]]
+  branch = "master"
+  name = "github.com/icrowley/fake"
+
+[[constraint]]
+  branch = "master"
+  name = "github.com/k0kubun/go-ansi"
+
+[[constraint]]
+  name = "github.com/schollz/progressbar"
+  version = "2.15.0"
+
+[[constraint]]
   name = "github.com/sirupsen/logrus"
   version = "1.4.2"
