+++ randomizer.go
@@ -3,10 +3,10 @@ package main
 import (
 	"errors"
 	"fmt"
+	"github.com/google/uuid"
 	"github.com/icrowley/fake"
 	"math"
 	"math/rand"
-	"os/exec"
 	"strconv"
 	"strings"
 	"time"
@@ -32,7 +32,8 @@ func RandomString(strlen int) string {
 // Random Number generator based on the min and max specified
 func RandomInt(min, max int) (int, error) {
 	if min >= max {
-		return 0, errors.New("max value is greater or equal to Min value, cannot generate data within these ranges")
+		return 0, errors.New("max value is greater or equal to Min value, " +
+			"cannot generate data within these ranges")
 	}
 	rand.Seed(time.Now().UnixNano())
 	return rand.Intn(max-min) + min, nil
@@ -65,7 +66,7 @@ func RandomFloat(min, max, precision int) (float64, error) {
 // Random calender date time generator
 func RandomCalenderDateTime(fromyear, toyear int) (time.Time, error) {
 	if fromyear > toyear {
-		return time.Now(), errors.New("Number of years behind is greater than number of years in future")
+		return time.Now(), errors.New("number of years behind is greater than number of years in future")
 	}
 	min := time.Now().AddDate(fromyear, 0, 0).Unix()
 	max := time.Now().AddDate(toyear, 0, 0).Unix()
@@ -122,11 +123,11 @@ func RandomTimetz(fromyear, toyear int) (string, error) {
 // Random bool generator based on if number is even or not
 func RandomBoolean() bool {
 	number, _ := RandomInt(1, 9999)
-	if number%2 == 0 {
-		return true
-	} else {
-		return false
+	var b bool
+	if b = false; number%2 == 0 {
+		b = true
 	}
+	return b
 }
 
 // Random Paragraphs
@@ -138,34 +139,27 @@ func RandomParagraphs() string {
 // Random IPv6 & IPv4 Address
 func RandomIP() string {
 	number, _ := RandomInt(1, 9999)
-	if number%2 == 0 {
-		return fake.IPv4()
-	} else {
-		return fake.IPv6()
+	var ip string
+	if ip = fake.IPv6(); number%2 == 0 {
+		ip = fake.IPv4()
 	}
+	return ip
 }
 
 // Random bit
 func RandomBit(max int) string {
 	var bitValue string
 	for i := 0; i < max; i++ {
-		if RandomBoolean() {
+		if bitValue = bitValue + "0"; RandomBoolean() {
 			bitValue = bitValue + "1"
-		} else {
-			bitValue = bitValue + "0"
 		}
 	}
 	return bitValue
 }
 
 // Random UUID
-func RandomUUID() (string, error) {
-	// To generate random UUID, we will use unix tool "uuidgen" (unix utility)
-	uuidString, err := exec.Command("uuidgen").Output()
-	if err != nil {
-		return "", fmt.Errorf("Unable to run uuidgen to generate UUID data: %v", err)
-	}
-	return strings.TrimSpace(string(uuidString)), nil
+func RandomUUID() string {
+	return strings.TrimSpace(uuid.New().String())
 }
 
 // Random Mac Address
@@ -202,30 +196,21 @@ func RandomTSVector() string {
 // Random Geometric data
 func RandomGeometricData(randomInt int, GeoMetry string, IsItArray bool) string {
 	var geometry []string
+	var data string
 	if GeoMetry == "point" { // Syntax for point datatype
-		if IsItArray { // If Array
-			return "\"(" + fake.DigitsN(2) + "," + fake.DigitsN(3) + ")\""
-		} else {
-			return "(" + fake.DigitsN(2) + "," + fake.DigitsN(3) + ")"
-		}
+		data = fmt.Sprintf("(%s,%s)", fake.DigitsN(2), fake.DigitsN(3))
+		return FormatForArray(data, IsItArray, true)
 	} else if GeoMetry == "circle" { // Syntax for circle datatype
-		if IsItArray { // If Array
-			return "\"(" + fake.DigitsN(2) + "," + fake.DigitsN(3) + ")," + fake.DigitsN(2) + ")\""
-		} else {
-			return "(" + fake.DigitsN(2) + "," + fake.DigitsN(3) + ")," + fake.DigitsN(2) + ")"
-		}
-
+		data = fmt.Sprintf("(%s,%s,%s)", fake.DigitsN(2), fake.DigitsN(3), fake.DigitsN(2))
+		return FormatForArray(data, IsItArray, true)
 	} else { // Syntax for the rest of geometry datatype
 		for i := 0; i < randomInt; i++ {
 			x, _ := RandomFloat(1, 10, 2)
 			y, _ := RandomFloat(1, 10, 2)
-			geometry = append(geometry, "("+fmt.Sprintf("%v", x)+","+fmt.Sprintf("%v", y)+")")
+			geometry = append(geometry, fmt.Sprintf("(%v,%v)", x, y))
 		}
-		if IsItArray { // If Array
-			return "\"(" + strings.Join(geometry, ",") + ")\""
-		} else {
-			return "(" + strings.Join(geometry, ",") + ")"
-		}
+		data = fmt.Sprintf("(%s)", strings.Join(geometry, ","))
+		return FormatForArray(data, IsItArray, true)
 	}
 	return ""
 }
@@ -240,71 +225,16 @@ func RandomLSN() string {
 func RandomTXID() string {
 	x, _ := strconv.Atoi(fake.DigitsN(8))
 	y, _ := strconv.Atoi(fake.DigitsN(8))
-	if x > y { // left side of ":" should be always less than right side
-		return fmt.Sprintf("%v:%v:", y, x)
-	} else {
-		return fmt.Sprintf("%v:%v:", x, y)
+	var z string
+	if z = fmt.Sprintf("%v:%v:", x, y); x > y { // left side of ":" should be always less than right side
+		z = fmt.Sprintf("%v:%v:", y, x)
 	}
-	return ""
+	return z
 }
 
 // Random JSON generator
 func RandomJson(IsItArray bool) string {
-	jsonData := `
-{
-    "_id": "%s",
-    "index": "%s",
-    "guid": "%s-%s-%s-%s-%s",
-    "isActive": "%s",
-    "balance": "%s.%s",
-    "website": "https://%s/%s",
-    "age": "%s",
-    "username": "%s",
-    "eyeColor": "%s",
-    "name": "%s",
-    "gender": "%s",
-    "company": "%s",
-    "email": "%s",
-    "phone": "%s",
-    "address": "%s",
-    "zipcode": "%s",
-    "state": "%s",
-    "country": "%s",
-    "about": "%s",
-    "Machine IP": "%s",
-    "job title": "%s",
-    "registered": "%s-%s-%sT%s:%s:%s-%s:%s",
-    "latitude": "%s.%s",
-    "longitude": "%s.%s",
-    "tags": [
-      "%s",
-      "%s",
-      "%s",
-      "%s",
-      "%s",
-      "%s",
-      "%s"
-    ],
-    "friends": [
-      {
-        "id": "%s",
-        "name": "%s"
-      },
-      {
-        "id": "%s",
-        "name": "%s"
-      },
-      {
-        "id": "%s",
-        "name": "%s"
-      }
-    ],
-    "greeting": "%s",
-    "favoriteBrand": "%s"
-  }
-`
-	jsonData = fmt.Sprintf(jsonData, RandomString(24), fake.DigitsN(10), RandomString(8),
-		RandomString(4), RandomString(4), RandomString(4), RandomString(12),
+	jsonData := fmt.Sprintf(JsonSkeleton(), RandomString(24), fake.DigitsN(10), RandomUUID(),
 		strconv.FormatBool(RandomBoolean()), fake.Digits(), fake.DigitsN(2), fake.DomainName(), fake.WordsN(1),
 		fake.DigitsN(2), fake.UserName(), fake.Color(), fake.FullName(), fake.Gender(), fake.Company(),
 		fake.EmailAddress(), fake.Phone(), fake.StreetAddress(), fake.Zip(), fake.State(), fake.Country(),
@@ -315,43 +245,14 @@ func RandomJson(IsItArray bool) string {
 		fake.WordsN(1), fake.WordsN(1), fake.WordsN(1), fake.WordsN(1), fake.DigitsN(2),
 		fake.FullName(), fake.DigitsN(2), fake.FullName(), fake.DigitsN(2), fake.FullName(), fake.Sentence(),
 		fake.Brand())
-
-	return FormatForArray(jsonData, IsItArray)
+	return FormatForArray(jsonData, IsItArray, false)
 }
 
 // Random XML Generator
 func RandomXML(IsItArray bool) string {
-	// XML Skeleton
-	xmlData := `
-<?xml version="1.0" encoding="UTF-8"?>
-<shiporder orderid="%s" xmlns:xsi="http://%s/%s/%s" xsi:noNamespaceSchemaLocation="shiporder.xsd">
-  <orderperson>%s</orderperson>
-  <shipto>
-    <name>%s</name>
-    <address>%s</address>
-    <city>%s</city>
-    <country>%s</country>
-    <email>%s</email>
-    <phone>%s</phone>
-  </shipto>
-  <item>
-    <title>%s</title>
-    <note>%s</note>
-    <quantity>%s</quantity>
-    <color>%s</color>
-    <price>%s.%s</price>
-  </item>
-  <item>
-    <title>%s</title>
-    <quantity>%s</quantity>
-    <price>%s.%s</price>
-  </item>
-</shiporder>
-`
-	xmlData = fmt.Sprintf(xmlData, fake.Digits(), fake.DomainName(),
+	xmlData := fmt.Sprintf(XMLSkeleton(), fake.Digits(), fake.DomainName(),
 		fake.DigitsN(4), fake.WordsN(1), fake.FullName(), fake.FullName(), fake.StreetAddress(), fake.City(),
 		fake.Country(), fake.EmailAddress(), fake.Phone(), fake.Title(), fake.Sentences(), fake.Digits(), fake.Color(),
 		fake.Digits(), fake.DigitsN(2), fake.Title(), fake.Digits(), fake.Digits(), fake.DigitsN(2))
-
-	return FormatForArray(xmlData, IsItArray)
+	return FormatForArray(xmlData, IsItArray, false)
 }
